requirements.txt
contourpy==1.1.0
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.2
mpmath==1.3.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sympy==1.12
tzdata==2023.3
